f1: change latitude and logitude type

+++ forsquare_api/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 class Location(models.Model):
     fsq_id = models.CharField(max_length=200)
-    latitude = models.FloatField()
-    longitude = models.FloatField()
+    latitude = models.DecimalField(max_digits=9, decimal_places=6)
+    longitude = models.DecimalField(max_digits=9, decimal_places=6)
     address = models.TextField()
     country = models.CharField(max_length=200)
     region = models.CharField(max_length=200)
@@ -12,3 +12,4 @@ class Location(models.Model):
 
     def __str__(self):
         return self.country + '-' + self.name
+
